feat: product & Menu 패키지 분리
