+++ README.md
@@ -244,8 +244,7 @@ docker compose -p kitchenpos up -d
 
 ### 속성
 
-- 주문은 주문 아이디, 주문 유형과 주문 상태, 주문 시각, 주문 항목 목록을 갖는다.
-    - 주문 아이디는 주문을 식별할 수 있다.
+- 주문은 주문 아이디주문 유형과 주문 상태, 주문 시각, 주문 항목 목록을 갖는다.
     - 주문 유형
         - 배달, 매장, 포장 주문이 존재한다. 이외의 주문 유형은 허용하지 않는다.
     - 주문 상태
