+++ src/test/java/kitchenpos/application/InMemoryProductRepository.java
@@ -1,7 +1,7 @@
 package kitchenpos.application;
 
-import kitchenpos.domain.Product;
-import kitchenpos.domain.ProductRepository;
+import kitchenpos.product.domain.Product;
+import kitchenpos.product.domain.ProductRepository;
 
 import java.util.*;
 import java.util.stream.Collectors;
@@ -28,8 +28,8 @@ public class InMemoryProductRepository implements ProductRepository {
     @Override
     public List<Product> findAllByIdIn(final List<UUID> ids) {
         return products.values()
-            .stream()
-            .filter(product -> ids.contains(product.getId()))
-            .collect(Collectors.toList());
+                .stream()
+                .filter(product -> ids.contains(product.getId()))
+                .collect(Collectors.toList());
     }
 }
